admin/debian/install-standard-packages.rb: A few minro tweaks.
Some support has been added for a bundle, but that has now moved to the invoking script and so will probably shortly be removed from here.

A few variables have been renamed from (say) _err to just _, to keep ruby 1.9 happy.

emacs24 has been changed to be plain emacs, so as to work on both wheezy and jessie systems.