admin/debian/install-advert-blocking-hosts-file.sh: Downlaod the latest advert-blocking hosts file and append it to the existing /etc/hosts.
This script should cope with being able to run multiple times. It will preserve the existing /etc/hosts file  until the line
containing the name Dan Pollock, then it will appened the hosts file that it manages to download.

It does *not* maintain any additions to the hosts file that are made after the downloaded hosts file addition.