Enhance Installer.rb to allow the caller to specify additional options which the caller then handles.
GetoptLong does not like to see options it has not been told about because it cannot guarantee parsing them properly.
When it sees such an option it issues a warning message and raises an exception. The former can be suppressed
using .quiet() and the latter can be caught. GetoptLong also modifies ARGV (probably using .shift()).
However, even if ARGV is restored, re-running GetoptLong does not work.

So to allow the caller the flexibility of additional arguments, accept an optional array of such arguments. Then,
if the base code sees an argument that it does not understand, it must have come from the caller. In that case,
if the caller has supplied a block, invoke it appropriately.

libs/ruby/Installer.rb: Allow caller to specify additional arguments which it may handle via a block. Add verbose as
a new basic option.

Note that the "opt.parse(&Proc.new())" trick is described here:

  http://mudge.name/2011/01/26/passing-blocks-in-ruby-without-block.html

The gist of that is:

"If Proc.new is called from inside a method without any arguments of its own, it will return a new Proc containing the block given to its surrounding method."

libs/ruby/Installer_spec.rb: Tweak the spec to test for --verbose.