Provide a script and underlying code to configure a system's hostname and IPv4 address.
There is not support yet for IPv6 addresses, nor for multiple IPv4 addresses.

libs/ruby/InstallHostnameAndAddress.rb: This module allows the hostname and (optionally) the IPv4 address to be configured.

admin/debian/configure-hostname-and-address.rb: This script parses arguments and the invokes the module appropriately.

admin/configuration.data: Add an entry for dns-extras: these are extrenal DNS servers to use after the primary one. dns-slaves is also renamed from dns-slave to emphasise that there may be more than one DNS slave system. (Currently no DNS slaves are configured so this value is not used in scripts yet).