libs/ruby/InstallDnsServer.rb: A script to handle instyaslling the environment for a DNS server.
Note that no configuration as such ios currently performed by this script. It is expected that
after installing the DNS server, a cronjob will be run regularly to update the DNS database as
required. Once that script is finalised, perhaps scheduling it via cron will be done as part
of the configuration here.