Fix hostname/IPv4 address configuration.
When this code was originally written, there was only one DNS zone file which provided host data.
That has changed, mostly to force all code to be able to cope and thereby avoid a tricky update years down the line.
So now the Host module needs to understand how to load the DNS zone files and collect all the hosts.
There is an assumption that hostnames are unique even across subnets.
The Host now includes the domain and netmask fields.

libs/ruby/Host.rb: Provide a way to find a host without the caller needing to load zone files.

libs/ruby/InstallHostnameAndAddress.rb: Restore the previously eliminated code.