Adjust pre-seeding code to allow for multiple lines of pre-seed.
Silently installing mysql-server via pre-seeding requires multiple lines to be fed to the debian pre-seeder.
Additionally, the package to install is not necessarily the *exactly* the first item on each line.

Fix the api to allow for those things and adjust the nis client installer to cope.

libs/ruby/Package.rb: Adjust the pre-seeding code to be more flexible ...

libs/ruby/InstallNisClient.rb: ... and update the sole existing user.