Add libs/ruby/Systems.rb to provide the basis for finding all reachable systems.
This will form part of a general systems health-check script.

admin/systems/192.168.1.flexbl.dns: Add "systype: host" to all systems considered to be hosts.

libs/ruby/Host.rb: Add support for systype.

libs/ruby/Systems.rb: A new module that provides support for determining all the known hosts in a domain or all the known hosts in a domain that are running SSH.