Begin to add support for makeself.
The idea is to produce a self-extracting archive that is simpler to transfer to the target system (one file rather than two).

Makefile: Add targets to build a makeself archive containing either a git bundle or the entire repo as a tree.

admin/standard-install-for-makeself.sh: For the moment this is just a placeholder stub.