A few further steps towards building a makeself .run file that can install the misc repo on a target system.
Makefile:

When building the bundle variant, make sure that standard-install.sh is available and not just in the bundle (since it is hard to extract from there without using git, which leads to a chicken and egg situation).

Pass in an argument of "bundle" or "archive" when building the .run script to make it easy for the pre-installer to decide what to do.

admin/standard-install.sh: Tweak some comments.

admin/standard-install-for-makeself.sh: First attempt at actually invoking the main installer shell script.