Makefile: Change the name of the target that builds the self-extracting archive.
As the self-extracting archive that contains a git bundle is named .run, rename the target that builds it "run".

The self-extracting archive that contains a tar file is also named .run, but that target is hardly used and not properly supported so no need to worry yet about whether that might cause potential confusion.