Make more rspec code run from outside its directory.
libs/ruby/Installer_spec.rb: Adjust $LOAD_PATH and the require to allow the rpsec source to load the code it will test (assuming it always lives in the same directory). In addition fix a bug when "--verbose" is the only thing specified on the command line. It should default to "--verbose --install --configure", and it does, but the test was wrong.