tools/scripts/download-dcu-from-eisner.py: Report a missing date as an accumulated error.
Reporting a missing date multiple times in the normal output stream leaves error text mixed in with expected text. Instead, report this along with other accumulated errors so that it appears at the end of the output. This makes it appear where errors expected to be found and where they can be removed from the expected text.