Allow a coin name to specify the JSON .id
The coins.txt file specified only a coin name, which corresponds to
.symbol in the JSON returned by coingecko.com. Unfortunately this can be
ambiguous (e.g. the symbol DOGE is used by binance-peg-dogecoin as well
as dogecoin). The simplest fix was just to drop the unwanted
"binance-peg-" variant but that became unsustainable as further coins
were added and further duplicates appeared.

Now coins.txt can optionally supply the expected JSON .id for a given
coin when disambiguation is necessary. This leaves the simple coin name
in all cases where disambiguation is not required (e.g. "FLOW") and the
new format can be used when multiple coins use the same .symbol (in this
case specify e.g. "DOGE:DOGECOIN") to pick the required coin.