Support Avalanche and display coin names
AVAX (Avalanche) is the only symbol (that I've found, so far) that is
specified in uppercase. It's awkward to handle just this case, and it
would become even more clumsy as more such symbols turn up. So require
the coins.txt file to specify the correct case for both the symbol and
the id.

To make the list of symbols and prices a shade more useful, add an extra
column that specifies the coin name, as recorded by coingecko.com.