Implement mediawiki back and restore scripts
mediawiki-backup.sh backups up everything needed to completely restore my mediawiki installation.
It includes the /var/lib/mediawiki tree, which should include all images and also all extensions, as well as the mediawiki configuration.
The script expects that automysqlbackup is running regularly.
borgbackup needs to be installed.

mediawiki-restore.sh restores a wiki that has been backup by the mediawiki-backup.sh script.
It modifies the LocalSettings.php script to set the wiki to read-only mode and also adjusts the wiki logo to make this clear. The intention is to run this script on an alternate system and have it perform a nightly restore so as to have a "hot standby" (read-only) copy of the wiki.