check-dec-cdrom-archive.py: Validate SHA256 file
Enable the code that checks the SHA256 checksum file.

The file is parsed assuming the hash is followed by a space and then either a space or an asterisk and finally a filename. Anything else is an error.

The SHA256 file must not be included inside itself (logically it cannot meaningfully be there).