tools/scripts/archive-links.py: Add history
Asking whether a link has already been archived is quicker than asking for a link to be archived, but avoiding asking is even quicker. Implement a history file to track URLs that have already been archived.
Include the date that the URL was archived to allow for the possibility or repeating archive requests if the last time a URL was archived was more than a specified number of days ago. This may never be implemented but allowing for it is cheap.